fix:fix method for change password

+++ isa-project/src/main/java/rs/ac/uns/ftn/isaproject/repository/users/UserAccountRepository.java
@@ -1,9 +1,10 @@
 package rs.ac.uns.ftn.isaproject.repository.users;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import rs.ac.uns.ftn.isaproject.model.users.UserAccount;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
-	UserAccount findByUsername( String username );
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
+	
+	UserAccount findByUsername(String username);
+
 }
